chore: clean-up ws init and oauth2
