+++ src/app/features/public/oauth2/components/oauth-callback/oauth-callback.component.ts
@@ -2,15 +2,13 @@ import { Component, OnInit, inject, DestroyRef } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { CommonModule } from '@angular/common';
 import { NzSpinModule } from 'ng-zorro-antd/spin';
-import {
-  OAuth2Service,
-  OAuth2LinkResponse,
-} from '../../../../shared/services/oauth2/oauth2.service';
+import { OAuth2Service } from '../../../../shared/services/oauth2/oauth2.service';
 import { HttpErrorResponse } from '@angular/common/http';
 import { AuthService } from '../../../../../shared/services/auth/auth.service';
 import { StorageService } from '../../../../../shared/services/auth/storage.service';
 import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
 import { LoginResponse } from '../../../../shared/models/auth/auth.model';
+import { OAuth2LinkResponse } from '../../../../shared/models/auth/oauth2.model';
 
 @Component({
   selector: 'app-oauth-callback',
