+++ src/app/features/shared/models/auth/oauth2.model.ts
@@ -1,0 +1,18 @@
+export interface OAuthCodeRequest {
+  code: string;
+  codeVerifier: string;
+}
+
+export interface LinkOAuthAccountRequest {
+  shouldLink: boolean;
+  provider?: string;
+  providerId?: string;
+  userId?: number;
+  password?: string;
+}
+
+export interface OAuth2LinkResponse {
+  providerName: string;
+  providerId: string;
+  userId: number;
+}
